backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and Caching
redis==5.0.1
hiredis==2.2.3

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# File Processing
python-magic==0.4.27
Pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.4

# Email
fastapi-mail==1.4.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# Search
elasticsearch==8.11.0

# AWS/S3
boto3==1.34.0
botocore==1.34.0

# Monitoring and Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# CORS
python-cors==1.0.0

# Environment
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
slugify==0.0.1
python-slugify==8.0.1
unidecode==1.3.7

# PDF Generation
reportlab==4.0.7
weasyprint==60.2

# Excel/CSV
xlsxwriter==3.1.9

# Validation
cerberus==1.3.5

# Cryptography
cryptography==41.0.8

# JSON Web Tokens
PyJWT==2.8.0

# HTTP Status
http-status==0.2.1

